+++ lib/views/home/login_background.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 
-import '../tools/arc_clipper.dart';
+import '../common_items/tools/arc_clipper.dart';
 
 class LoginBackground extends StatelessWidget {
   Widget _topHalf(BuildContext context) {
